requirements.txt
fastapi==0.128.0
joblib==1.4.2
pandas==2.3.3
pydantic==2.12.5
python-dotenv==1.2.1
Requests==2.32.5
scikit_learn==1.8.0
streamlit==1.44.1
uvicorn==0.40.0
